% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_error}
\alias{capture_error}
\title{Capture error in a GET response}
\usage{
capture_error(response)
}
\arguments{
\item{response}{of a GET functions}
}
\description{
Capture error in a GET response
}
