% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_density_map.R
\name{gen_density_map}
\alias{gen_density_map}
\title{Generate density maps from flight data}
\usage{
gen_density_map(flight_data, crs = "+init=epsg:4326", to = "raster",
  min_value = 0, from = "lines")
}
\arguments{
\item{flight_data}{Dataframe with data from flights.}

\item{crs}{A character string of projection arguments}

\item{to}{raster-> export data in raster format}

\item{min_value}{Every value under this parameter has not be showed in raster output.
Only works when from="points"}

\item{from}{A character strint to select the density calculation font. It
could be "lines" or "points"}
}
\value{
a density map in different type formats.Right now it is only developed "raster" option.
}
\description{
Allow to generate density maps from flight data in differents output types
}
\examples{
raster<-gen_density_map(flight_data,min_value=0,from="points")
raster<-gen_density_map(flight_data,from="lines")
}

