% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airport_data.R
\name{get_airport_data}
\alias{get_airport_data}
\title{get_airport_data}
\usage{
get_airport_data(
  username = NULL,
  password = NULL,
  airport = NULL,
  begin = NULL,
  end = NULL,
  option = c("departures", "arrivals")
)
}
\arguments{
\item{username}{Your 'OpenSky Network' username.}

\item{password}{Your 'OpenSky Network' password.}

\item{airport}{ICAO identifier for the airport.}

\item{begin}{Start of time interval to retrieve flights for as Unix
time (seconds since epoch).}

\item{end}{End of time interval to retrieve flights for as Unix time
(seconds since epoch).}

\item{option}{String indicating whether you want to request the departure
or arrival flights. "arrivals" or "departures".}
}
\value{
A dataframe with the list of flights arriving to the airport defined
in airport in the period specified between begin and end.
}
\description{
Retrieve flights for a certain airport which arrived within
a given time interval. You should specified if you want departure or
arrival flights in the function.
}
\examples{
\dontrun{get_airport_data("your_username", "your_password", airport = "EDDF",
 begin = 1517227200, end = 1517230800, option = "arrivals")}
  \dontrun{get_airport_data("your_username", "your_password", airport = "EDDF",
 begin = 1517227200, end = 1517230800, option = "departures")}

}
