% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_data.R
\name{get_track_data}
\alias{get_track_data}
\title{get_track_data}
\usage{
get_track_data(username = NULL, password = NULL, icao24 = NULL, time = NULL)
}
\arguments{
\item{username}{Your 'OpenSky Network' username.}

\item{password}{Your 'OpenSky Network' password.}

\item{icao24}{Unique ICAO 24-bit address of the transponder in hex string
representation. All letters need to be lower case}

\item{time}{Unix time in seconds since epoch. It can be any time between
start and end of a known flight. If time = 0, get the live track if there is any flight ongoing for the given aircraft.}
}
\value{
A dataframe with the trajectory of a certain aircraft at a given
time.
}
\description{
Retrieve flights for a certain airport which arrived within a
given time interval. If no flights are found for the given period, HTTP
stats 404 - Not found is returned with an empty response body.
}
\examples{
\dontrun{get_track_data("luisgasco", "masterEOI2015", icao24 = 3c4b26, time = 0)}

}
