% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOS_anom.R
\name{getOS_anom}
\alias{getOS_anom}
\title{Get OpenSky data anonymously}
\usage{
getOS_anom(complete = TRUE, by = 6:8)
}
\arguments{
\item{complete}{TRUE if you want to get complete cases depending on "by" variable.}

\item{by}{A vector of numbers that specify the columns to take into account to complete cases.
By default the filter works on Latitude, Longitude and Altitude column}
}
\value{
A dataframe with every flight registered on OpenSky Network in real-time.
}
\description{
Allow to access OpenSky Network API without authentication. There are some
limitations:
\itemize{
\item Anonymous users can only get the most recent state vectors
\item Anonymous users can only retrieve data with a time resultion of 10 seconds
}
\href{https://opensky-network.org/apidoc/rest.html#limitiations-for-anonymous-unauthenticated-users}{More API information}
}
\examples{
data_completecases <- getOS_anom(TRUE,by=6:7)
data <- getOS_anom()
}

