% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOS_auth.R
\name{getOS_auth}
\alias{getOS_auth}
\title{Get OpenSky data authenticated}
\usage{
getOS_auth(user = as.character(), key = as.character(), complete = TRUE,
  by = 6:8)
}
\arguments{
\item{user}{A valid user of OpenSky Network .}

\item{key}{The password of the user specified before.}

\item{complete}{TRUE if you want to get complete cases depending on "by" variable.}

\item{by}{A vector of numbers that specify the columns to take into account to complete cases.
By default the filter works on Latitude, Longitude and Altitude column}
}
\value{
A dataframe with every flight registered on OpenSky Network in real-time.
}
\description{
Allow to access OpenSky Network API using your valid OpenSky account,
The rate limitations are:
\itemize{
\item OpenSky users can retrieve data of up to 1 hour in the past
\item OpenSky users can retrieve data with a time resultion of 5 seconds.
}
\href{https://opensky-network.org/apidoc/rest.html#limitiations-for-anonymous-unauthenticated-users}{More API information}
}
\examples{
data_completecases <- getOS_auth("your_user","your_password",TRUE,by=6:7)
data <- getOS_auth("your_user","your_password")
}

