% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOS_own.R
\name{getOS_own}
\alias{getOS_own}
\title{Get your sensors' data}
\usage{
getOS_own(user = as.character(), key = as.character(), complete = TRUE,
  by = 6:8, serials = as.vector())
}
\arguments{
\item{user}{A valid user of OpenSky Network .}

\item{key}{The password of the user specified before.}

\item{complete}{TRUE if you want to get complete cases depending on "by" variable.}

\item{by}{A vector of numbers that specify the columns to take into account to complete cases.
By default the filter works on Latitude, Longitude and Altitude column}

\item{serials}{A character vector with the serials codes of your receivers you want to retrieve}
}
\value{
A dataframe with every flight registered on OpenSky Network in real-time, but only related
to your own sensors registered in the platform.
}
\description{
Allow to access to vectors for your own sensors without any rate limitation.Note that authentication is required for this operation, otherwise you will
get a 403 - Forbidden.
\href{https://opensky-network.org/apidoc/rest.html#limitiations-for-anonymous-unauthenticated-users}{More API information}
}
\examples{
your_own_data <- getOS_own("your_user","your_password",
serials=c("your_sensor_serial1","your_sensor_serial2"))
your_own_data2 <- getOS_own("your_user","your_password",
serials="your_sensor_serial1")
your_own_data3 <- getOS_own("your_user","your_password")
}

