% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openskyr.R
\docType{package}
\name{openskyr}
\alias{openskyr}
\title{openskyr: A third-party client to facilitate the access to the data offered by
'OpenSky Network' in its REST API.}
\description{
DESCRIPTION
}
\section{openskyr functions}{

The openskyr functions allow you to access of any of the REST-API
functions available in 'OpenSky Network'. In many of the requests,
you would need a user account that you can create in their website.
}

\seealso{
Useful links:
\itemize{
\item 'OpenSky Network' website \url{https://opensky-network.org}
\item API Rest docs \url{https://opensky-network.org/apidoc/rest.html}
\item Report bugs at \url{https://github.com/luisgasco/openskyr/issues}
}
}
\author{
\strong{Maintainer}: Luis Gasco Sanchez \href{mailto:luisgascosanchez@gmail.com}{luisgascosanchez@gmail.com}
}
