% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_pl.R
\name{trans_pl}
\alias{trans_pl}
\title{Transform point trajectories to lines}
\usage{
trans_pl(data, long, lat, id = NULL)
}
\arguments{
\item{data}{Dataframe with data from flights.}

\item{long}{A character string of longitude parameter}

\item{lat}{A character string of latitude parameter}

\item{id}{A character string of id of plane, icao24.}
}
\value{
SpatialLines object of line trajectories followed by flights
}
\description{
Allow to transform a point data dataframe of flight path to SpatialLines data
}
\examples{
traj <- trans_pl(data = points_df,
                   long = "longitude",
                   lat = "latitude",
                   id="icao24")
}

