% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airport_data.R
\docType{data}
\name{get_airport_data}
\alias{get_airport_data}
\alias{col_airport_names}
\title{get_airport_data}
\format{
An object of class \code{character} of length 12.
}
\usage{
col_airport_names
}
\arguments{
\item{username}{Your opensky-network username.}

\item{password}{Your opensky-network password.}

\item{airport}{ICAO identier for the airport.}

\item{begin}{Start of time interval to retrieve flights for as Unix time (seconds since epoch).}

\item{end}{End of time interval to retrieve flights for as Unix time (seconds since epoch).}

\item{option}{String indicating whether you want to request the departure or arrival flights. "arrivals" or "departures".}
}
\value{
A dataframe with the list of flights arriving to the airport defined in airport in the period specified between begin and end.
}
\description{
Retrieve flights for a certain airport which arrived within a given time interval.
You should specified if you want departure or arrival flights in the function.
}
\keyword{datasets}
