% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_vectors.R
\docType{data}
\name{get_state_vectors}
\alias{get_state_vectors}
\alias{col_statevectors_names}
\title{get_state_vectors}
\format{
An object of class \code{character} of length 16.
}
\usage{
col_statevectors_names
}
\arguments{
\item{username}{Your opensky-network username.}

\item{password}{Your opensky-network password.}

\item{time}{(optional) The time in seconds since epoch (Unix time stamp to retrieve states for. Current time will be used if omitted.}

\item{icao24}{(optional) One or more ICAO24 transponder addresses represented by a hex string (e.g. abc9f3). To filter multiple ICAO24 append the property once for each address. If omitted, the state vectors of all aircraft are returned.}

\item{lamin}{(optional) Lower bound for the latitude in decimal degrees.}

\item{lomin}{(optional) Lower bound for the longitude in decimal degrees.}

\item{lamax}{(optional) Upper bound for the latitude in decimal degrees.}

\item{lomax}{(optional) Upper bound for the longitude in decimal degrees.}
}
\value{
A data.frame with fields specified in the opensky-network REST API \href{https://opensky-network.org/apidoc/rest.html#response}{documentation}.
}
\description{
Retrieve any state vector at the OpenSky-Network API. Depending on whether you use your personal credentials or not,
you will have some limitations rate limits as explained in \href{https://opensky-network.org/apidoc/rest.html#limitations}{opensky-network documentation}
}
\keyword{datasets}
