% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openskyr.R
\docType{package}
\name{openskyR}
\alias{openskyR}
\title{openskyr: A package for computating the notorious bar statistic.}
\description{
DESCRIPTION
}
\section{openskyR functions}{

The openskyR functions ...
}

\seealso{
Useful links:
\itemize{
\item Opensky-network website https://opensky-network.org
\item API Rest docs https://opensky-network.org/apidoc/rest.html
\item Report bugs at GITUB CODE
}
}
\author{
\strong{Maintainer}: Luis Gasco Sanchez \href{mailto:luisgascosanchez@gmail.com}{luisgascosanchez@gmail.com}
}
