% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recover_names}
\alias{recover_names}
\title{Recover column names}
\usage{
recover_names(option)
}
\arguments{
\item{option}{Option of column names recovered}
}
\description{
Recover column names
}
\keyword{internal}
